% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_setup.R
\name{add_chapter}
\alias{add_chapter}
\title{Add a new .qmd file representing a Quarto book chapter}
\usage{
add_chapter(filename = here("chapters", "chapter_0X.qmd"))
}
\arguments{
\item{filename}{Character indicating the path to the .qmd file to
be created.}
}
\value{
Character with path to .qmd files.
}
\description{
Add a new .qmd file representing a Quarto book chapter
}
\examples{
path <- tempdir()
c <- add_chapter(filename = tempfile(fileext = ".qmd"))
}
