% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project_tree}
\alias{create_project_tree}
\title{Create project tree}
\usage{
create_project_tree(rootdir = here::here(), quarto_book = TRUE)
}
\arguments{
\item{rootdir}{Path to root directory where the structure will be created.
Default: \code{here::here()}.}

\item{quarto_book}{Logical indicating whether to set up a Quarto book.
Default: TRUE.}
}
\value{
A character vector with paths to the directory structure.
}
\description{
This function creates a standard directory structure for scientific projects.
}
\examples{
rootdir <- file.path(tempdir(), paste0(sample(1:100, 1), "_tidyproj_ex"))
c <- create_project_tree(rootdir)
}
