% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_rmd}
\alias{create_rmd}
\title{Create an .Rmd file for code}
\usage{
create_rmd(rootdir = here::here(), filename = "00_new_code_file.Rmd")
}
\arguments{
\item{rootdir}{Path to root of the directory created
with \code{create_project_tree()}. Default: \code{here::here().}
Default: \code{here::here()}.}

\item{filename}{Character specifying the name of the .Rmd file to be
created. Default: 0_new_code_file.Rmd.}
}
\value{
Path to newly created .Rmd file.
}
\description{
Create an .Rmd file for code
}
\examples{
rootdir <- file.path(tempdir(), "tidyproj_example")
c <- create_project_tree(rootdir)

# Create file
file <- create_rmd(rootdir, filename = "05_new_analysis.Rmd")
}
