% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_setup.R
\name{setup_quarto_book}
\alias{setup_quarto_book}
\title{Set up directory for a Quarto book}
\usage{
setup_quarto_book(rootdir = here())
}
\arguments{
\item{rootdir}{Character with path to the root directory}
}
\value{
A character vector with paths to the directory structure.
}
\description{
Set up directory for a Quarto book
}
\examples{
rootdir <- tempdir()

setup_quarto_book(rootdir)
}
