% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project_tree}
\alias{create_project_tree}
\title{Create project tree}
\usage{
create_project_tree(create_renv = TRUE)
}
\arguments{
\item{create_renv}{Logical indicating whether to lockfile to handle
package dependencies with the package {renv}. Default: TRUE.}
}
\value{
A standard directory structure
}
\description{
This function creates a standard directory structure for scientific projects.
}
